Create PasswordGenerator.java
Just the basic logic outline. Works flawlessly as I tested it. To check, create the PasswordGenerator.class file and you should be good to go. Make sure you space out a bit for adding this to the main project. Remember, people cant get access to the task manager unless they create an account, and to create an account, they need a password, hence this generator comes into play. Just basic and simple logic of if statements tbh. 

Heres the method on how  to run it:
(In your terminal window)

1) Javac PasswordGenerator.java

2) Java PasswordGenerator

3) Follow along and type whatever you are prompted with.

Tell me if you find any errors or anything for improvement, obviously its still in development.

Signed-off-by: Ifhezu <90856703+Ifhezu@users.noreply.github.com>